支持ICM42688与MMC5603

+++ hw/bsp/c-mini/drivers/bsp_time.h
@@ -8,9 +8,9 @@ extern "C" {
 
 uint32_t bsp_time_get_ms();
 
-uint32_t bsp_time_get_us();
+uint64_t bsp_time_get_us();
 
-float bsp_time_get();
+uint64_t bsp_time_get();
 
 #ifdef __cplusplus
 }
