+++ src/module/engineer_chassis/mod_engineer_chassis.hpp
@@ -89,9 +89,9 @@ class Chassis {
 
   float dt_ = 0.0f;
 
-  float last_wakeup_ = 0.0f;
+  uint64_t last_wakeup_ = 0;
 
-  float now_ = 0.0f;
+  uint64_t now_ = 0;
 
   RefForChassis ref_;
